+++ src/main/java/org/manuel/mysportfolio/config/mvc/WebSecurityConfig.java
@@ -31,7 +31,7 @@ public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
   }
 
   @Override
-  public void configure(WebSecurity web) {
+  public void configure(final WebSecurity web) {
     web.ignoring().antMatchers("/v2/api-docs",
         "/configuration/ui",
         "/swagger-resources/**",