+++ src/main/java/org/manuel/mysportfolio/model/dtos/match/MatchDto.java
@@ -4,6 +4,9 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonPOJOBuilder;
+import io.github.manuelarte.spring.manuelartevalidation.constraints.groups.New;
+import io.github.manuelarte.spring.manuelartevalidation.constraints.groups.PartialUpdate;
+import io.github.manuelarte.spring.manuelartevalidation.constraints.groups.Update;
 import java.time.Instant;
 import java.util.List;
 import java.util.Map;
@@ -19,9 +22,6 @@ import org.manuel.mysportfolio.model.dtos.team.TeamTypeDto;
 import org.manuel.mysportfolio.model.entities.Place;
 import org.manuel.mysportfolio.model.entities.TeamOption;
 import org.manuel.mysportfolio.model.entities.match.type.MatchType;
-import org.manuel.mysportfolio.validation.NewEntity;
-import org.manuel.mysportfolio.validation.PartialUpdateEntity;
-import org.manuel.mysportfolio.validation.UpdateEntity;
 
 @JsonDeserialize(builder = MatchDto.MatchDtoBuilder.class)
 @JsonInclude(JsonInclude.Include.NON_NULL)
@@ -30,14 +30,14 @@ import org.manuel.mysportfolio.validation.UpdateEntity;
 @lombok.Builder(toBuilder = true)
 public class MatchDto<H extends TeamTypeDto, A extends TeamTypeDto> {
 
-  @Null(groups = {NewEntity.class, UpdateEntity.class, PartialUpdateEntity.class})
+  @Null(groups = {New.class, Update.class, PartialUpdate.class})
   private final String id;
 
-  @Null(groups = NewEntity.class)
-  @NotNull(groups = {UpdateEntity.class, PartialUpdateEntity.class})
+  @Null(groups = New.class)
+  @NotNull(groups = {Update.class, PartialUpdate.class})
   private final Long version;
 
-  @NotNull(groups = {NewEntity.class, UpdateEntity.class})
+  @NotNull(groups = {New.class, Update.class})
   private final MatchType type;
 
   @NotNull
@@ -66,7 +66,7 @@ public class MatchDto<H extends TeamTypeDto, A extends TeamTypeDto> {
       @NotEmpty
           String> chips;
 
-  @Null(groups = {NewEntity.class, UpdateEntity.class, PartialUpdateEntity.class})
+  @Null(groups = {New.class, Update.class, PartialUpdate.class})
   private final String createdBy;
 
   public static <H extends TeamTypeDto, A extends TeamTypeDto> MatchDtoBuilder builder() {