+++ src/main/java/org/manuel/mysportfolio/model/dtos/playerprofile/PlayerProfileSportSeasonSummaryDto.java
@@ -3,11 +3,11 @@ package org.manuel.mysportfolio.model.dtos.playerprofile;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonPOJOBuilder;
+import io.github.manuelarte.spring.manuelartevalidation.constraints.groups.New;
+import io.github.manuelarte.spring.manuelartevalidation.constraints.groups.PartialUpdate;
+import io.github.manuelarte.spring.manuelartevalidation.constraints.groups.Update;
 import java.math.BigDecimal;
 import javax.validation.constraints.Null;
-import org.manuel.mysportfolio.validation.NewEntity;
-import org.manuel.mysportfolio.validation.PartialUpdateEntity;
-import org.manuel.mysportfolio.validation.UpdateEntity;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonDeserialize(builder = PlayerProfileSportSeasonSummaryDto.PlayerProfileSportSeasonSummaryDtoBuilder.class)
@@ -16,16 +16,16 @@ import org.manuel.mysportfolio.validation.UpdateEntity;
 @lombok.Builder
 public class PlayerProfileSportSeasonSummaryDto {
 
-  @Null(groups = {NewEntity.class, UpdateEntity.class, PartialUpdateEntity.class})
+  @Null(groups = {New.class, Update.class, PartialUpdate.class})
   private final Integer numberOfMatchesPlayed;
 
-  @Null(groups = {NewEntity.class, UpdateEntity.class, PartialUpdateEntity.class})
+  @Null(groups = {New.class, Update.class, PartialUpdate.class})
   private final Integer numberOfGoals;
 
-  @Null(groups = {NewEntity.class, UpdateEntity.class, PartialUpdateEntity.class})
+  @Null(groups = {New.class, Update.class, PartialUpdate.class})
   private final BigDecimal averageGoalRate;
 
-  @Null(groups = {NewEntity.class, UpdateEntity.class, PartialUpdateEntity.class})
+  @Null(groups = {New.class, Update.class, PartialUpdate.class})
   private final Integer numberOfAssists;
 
   @JsonPOJOBuilder(withPrefix = "")