+++ src/main/java/org/manuel/mysportfolio/model/dtos/teamtousers/UserInTeamDto.java
@@ -3,6 +3,8 @@ package org.manuel.mysportfolio.model.dtos.teamtousers;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonPOJOBuilder;
+import io.github.manuelarte.spring.manuelartevalidation.constraints.groups.New;
+import io.github.manuelarte.spring.manuelartevalidation.constraints.groups.Update;
 import java.time.LocalDate;
 import java.util.Set;
 import javax.validation.constraints.AssertTrue;
@@ -10,8 +12,6 @@ import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.PastOrPresent;
 import org.manuel.mysportfolio.model.entities.teamtouser.UserInTeam;
-import org.manuel.mysportfolio.validation.NewEntity;
-import org.manuel.mysportfolio.validation.UpdateEntity;
 
 @JsonDeserialize(builder = UserInTeamDto.UserInTeamDtoBuilder.class)
 @JsonInclude(JsonInclude.Include.NON_NULL)
@@ -20,13 +20,13 @@ import org.manuel.mysportfolio.validation.UpdateEntity;
 @lombok.Builder(toBuilder = true)
 public class UserInTeamDto {
 
-  @NotNull(groups = {NewEntity.class, UpdateEntity.class})
+  @NotNull(groups = {New.class, Update.class})
   @PastOrPresent
   private final LocalDate from;
 
   private final LocalDate to;
 
-  @NotEmpty(groups = {NewEntity.class, UpdateEntity.class})
+  @NotEmpty(groups = {New.class, Update.class})
   @lombok.Singular
   private final Set<UserInTeam.UserInTeamRole> roles;