+++ src/main/java/org/manuel/mysportfolio/validation/UpdateEntity.java
@@ -1,5 +1,0 @@
-package org.manuel.mysportfolio.validation;
-
-public interface UpdateEntity {
-
-}